improvement: support on_delete: :nilify for specific columns
Introduce the possibility to specify which columns of a foreign key
should be set to NULL when the referenced resource is deleted.
This provides fine-grained control over the `on_delete: :nilify`
foreign key constraint, where the default behavior is set all columns of
the foreign key to NULL.

This is useful for example in cases where multitenancy is employed on
resources, and a resource A has a `(tenant_id, resource_b_id)` foreign
key that references the `tenant_id` and `id` columns of resource B: when
the resource B is deleted, we'd like to nilify only the `resource_b_id`
column of resource A and not its `tenant_id`.

Note: this functionality is supported by Ecto's `on_delete: {:nilify,
columns}` option and ultimately by Postgres, although only from v15.0
onwards.

For details see:
https://www.postgresql.org/docs/current/ddl-constraints.html#DDL-CONSTRAINTS-FK
https://hexdocs.pm/ecto_sql/Ecto.Migration.html#references/2-options

Signed-off-by: Davide Briani <davide.briani@secomind.com>