fix: self-join if combination queries require more fields
In a query like this:

```elixir
Post
|> Ash.Query.combination_of([
  Ash.Query.Combination.base(
    filter: expr(score == 10),
    select: [:id, :score],
  ),
  Ash.Query.Combination.union_all(
    filter: expr(score == 20),
    select: [:id, :score],
  )
])
|> Ash.Query.filter(category == "category1")
|> Ash.read!()
```

you'd get a SQL query something like this (before my fix):

```sql
SELECT id, score FROM (
  SELECT id, score
  FROM posts
  WHERE score = 10
  UNION ALL
  SELECT id, score
  FROM posts
  WHERE score = 20
) AS posts
WHERE posts.category = "category1"
```

Which wouldn't work because `category` is not a part of the
inner selects. Now we add a `JOIN` on primary key to the table,
and then include all the fields we're missing from that table,
in any conditions that we would reference a field that wasn't
selected in the combinations.