fix(#582) Use new attribute source in down migration
When an attribute is renamed, alongside other changes to it (like changing default or not null constraint), the generated up migration first renames the column, then modifies it.

For the generated down migration, the order of operations is reversed, so we need to use the new column name when modifying the column, before renaming it back.