test: Add a test for an expr calc using an aggregation in a sort
This replicates an issue seen.

We have an expression calculation using a aggregation from a transitive
relationship.  If that calculation is used in a sort, the text of the
expression is passed as the aggregate name, instead of a variable name
such as "aggregate0" etc.

E.g.
```
* ** (RuntimeError) Unbound aggregate field:
"first(comments[field: :created_at, query: [filter: author_id ==
\"54c6b3da-d3a5-41f6-8409-780fed9fe184\"]])"
```